Adds new BaseCollection class without statistics.
The existing Collection class always generates statistsical aggregations this
extracts separates that out and lets user instantiate a BaseCollection for
searching collections without the statistics.
